requirements.txt
asgiref==3.2.10
azure-common==1.1.26
azure-core==1.11.0
azure-cosmos==4.2.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
isodate==0.6.0
ms-identity-web @ git+https://github.com/azure-samples/ms-identity-python-samples-common@9b32cef36c6c0b73c1b3237fdd66128d89f90a62
msal==1.10.0
msrest==0.6.21
oauthlib==3.1.0
Pillow==8.1.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
